simplifying processing est_tp line

+++ rateman/station.py
@@ -77,7 +77,7 @@ class Station:
         return self._loop
 
     @property
-    def expected_throughput(self) -> int:
+    def expected_throughput(self) -> float:
         """
         Return the current estimated throughput.
         """
